refactor: 상수 분리
