+++ constants/admin.ts
@@ -1,0 +1,7 @@
+const ADMIN_CATEGORY_OBJECT = [
+  { id: 0, name: '상품 관리', href: '/admin/product' },
+  { id: 1, name: '회원 관리', href: '/admin/user' },
+];
+const ADMIN_CATEGORY_ARRAY = ['/admin/product', '/admin/user'];
+
+export { ADMIN_CATEGORY_OBJECT, ADMIN_CATEGORY_ARRAY };
