+++ constants/server.ts
@@ -1,0 +1,3 @@
+const BASE_URL = 'https://howabout.site';
+
+export { BASE_URL };
