+++ constants/style.ts
@@ -1,0 +1,20 @@
+const representative = {
+  color: '#9d7446',
+  maxWidth: '1420px',
+};
+const modalStyle = {
+  overlay: {
+    backgroundColor: 'rgba(0,0,0,0.5)',
+  },
+  content: {
+    left: '0',
+    margin: 'auto',
+    width: '100%',
+    maxWidth: '500px',
+    height: '500px',
+    padding: '0',
+    overflow: 'hidden',
+  },
+};
+
+export { representative, modalStyle };
