+++ contexts/color.ts
@@ -1,0 +1,5 @@
+import { createContext } from 'react';
+
+const ColorContext = createContext({ color: 'oragne' });
+
+export default ColorContext;
